kobase: improve multi-platform support
Use `${TARGETPLATFORM}` rather than relying on `$(uname -m)` for
detecting the target architecture when building the image (as the later
can be wrong, e.g. when building form ARMhf with Github Actions using a
Aarch64 runner).

This also add support for ARMhf, and fix the shfmt install (which wasn't
taking into account the target platform).